+++ history.md
@@ -1,3 +1,8 @@
+# 2022-06-04
+
+- Add configuration `workspace.openOutputCommand`.
+- Log channel message of vim when `g:node_client_debug` enabled.
+
 # 2022-05-30
 
 - Disable `progressOnInitialization` for language client by default.